+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [Unreleased]
+
+## Changed
+
+- [TD-5468] Change the resource_acl_path to acl_path for kong update version
+
 ## [5.10.0] 2023-07-06
 
 ## Changed